+++ devices/modes-panel.js
@@ -25,6 +25,7 @@ class ModesPanel extends AlarmDevice {
         // Publish discovery message
         if (!this.discoveryData.length) { await this.initDiscoveryData() }
         await this.publishDiscoveryData()
+        await this.online()
 
         // This is a polled device so don't use common publish/subscribe function
         if (this.subscribed) {
@@ -43,7 +44,6 @@ class ModesPanel extends AlarmDevice {
             // Mark device as subscribed
             this.subscribed = true
         }
-        this.online()
     }
 
     initDiscoveryData() {
@@ -116,14 +116,14 @@ class ModesPanel extends AlarmDevice {
     async trySetMode(message, delay) {
         await utils.sleep(delay)
         let targetMode
-        switch(message) {
-            case 'DISARM':
+        switch(message.toLowerCase()) {
+            case 'disarm':
                 targetMode = 'disarmed'
                 break
-            case 'ARM_HOME':
+            case 'arm_home':
                 targetMode = 'home'
                 break
-            case 'ARM_AWAY':
+            case 'arm_away':
                 targetMode = 'away'
                 break
             default: